+++ Common/Database/Tables/Items.cs
@@ -5,14 +5,14 @@ namespace Common.Database.Tables
 {
     public class SubRequire
     {
-        public int Level { get; set; }
-        public int Skill { get; set; }
-        public int SkillRank { get; set; }
-        public int Spell { get; set; }
-        public int HonorRank { get; set; }
-        public int CityRank { get; set; }
-        public int ReputationFaction { get; set; }
-        public int ReputationRank { get; set; }
+        public int Level { get; set; } = 0;
+        public int Skill { get; set; } = 0;
+        public int SkillRank { get; set; } = 0;
+        public int Spell { get; set; } = 0;
+        public int HonorRank { get; set; } = 0;
+        public int CityRank { get; set; } = 0;
+        public int ReputationFaction { get; set; } = 0;
+        public int ReputationRank { get; set; } = 0;
     }
 
     /// <summary>
@@ -36,13 +36,13 @@ namespace Common.Database.Tables
 
     public class SubResistence
     {
-        public int Armor { get; set; }
-        public int Holy { get; set; }
-        public int Fire { get; set; }
-        public int Nature { get; set; }
-        public int Frost { get; set; }
-        public int Shadow { get; set; }
-        public int Arcane { get; set; }
+        public int Armor { get; set; } = 0;
+        public int Holy { get; set; } = 0;
+        public int Fire { get; set; } = 0;
+        public int Nature { get; set; } = 0;
+        public int Frost { get; set; } = 0;
+        public int Shadow { get; set; } = 0;
+        public int Arcane { get; set; } = 0;
     }
 
     /// <summary>
@@ -63,20 +63,20 @@ namespace Common.Database.Tables
     {
         public ObjectId Id { get; set; }
 
-        public int Entry { get; set; }
-        public int Class { get; set; }
-        public int SubClass { get; set; }
-        public string Name { get; set; }
-        public int DisplayId { get; set; }
-        public int Quality { get; set; }
-        public int Flags { get; set; }
-        public int BuyCount { get; set; }
-        public int BuyPrice { get; set; }
-        public int SellPrice { get; set; }
-        public int InventoryType { get; set; }
-        public int AllowableClass { get; set; }
-        public int AllowableRace { get; set; }
-        public int ItemLevel { get; set; }
+        public int Entry { get; set; } = 0;
+        public int Class { get; set; } = 0;
+        public int SubClass { get; set; } = 0;
+        public string Name { get; set; } = "NotFound";
+        public int DisplayId { get; set; } = 0;
+        public int Quality { get; set; } = 0;
+        public int Flags { get; set; } = 0;
+        public int BuyCount { get; set; } = 0;
+        public int BuyPrice { get; set; } = 0;
+        public int SellPrice { get; set; } = 0;
+        public int InventoryType { get; set; } = 0;
+        public int AllowableClass { get; set; } = 0;
+        public int AllowableRace { get; set; } = 0;
+        public int ItemLevel { get; set; } = 0;
 
         public SubRequire SubRequired { get; set; }
         public List<SubStat> SubStats { get; set; }
@@ -84,36 +84,36 @@ namespace Common.Database.Tables
         public SubResistence SubResistences { get; set; }
         public List<SubItemSpell> SubSpells { get; set; }
 
-        public int MaxCount { get; set; }
-        public int Stackable { get; set; }
-        public int ContainerSlots { get; set; }
+        public int MaxCount { get; set; } = 0;
+        public int Stackable { get; set; } = 0;
+        public int ContainerSlots { get; set; } = 0;
 
-        public int Delay { get; set; }
-        public int AmmoType { get; set; }
-        public float RangedModRange { get; set; }
+        public int Delay { get; set; } = 0;
+        public int AmmoType { get; set; } = 0;
+        public float RangedModRange { get; set; } = 0;
 
-        public int Bonding { get; set; }
+        public int Bonding { get; set; } = 0;
         public string Description { get; set; }
-        public int PageText { get; set; }
-        public int LanguageId { get; set; }
-        public int PageMaterial { get; set; }
-        public int StartQuest { get; set; }
-        public int LockId { get; set; }
-        public int Material { get; set; }
-        public int Sheath { get; set; }
-        public int RandomProperty { get; set; }
-        public int Block { get; set; }
-        public int ItemSet { get; set; }
-        public int MaxDurability { get; set; }
-        public int Area { get; set; }
-        public int Map { get; set; }
-        public int BagFamily { get; set; }
+        public int PageText { get; set; } = 0;
+        public int LanguageId { get; set; } = 0;
+        public int PageMaterial { get; set; } = 0;
+        public int StartQuest { get; set; } = 0;
+        public int LockId { get; set; } = 0;
+        public int Material { get; set; } = 0;
+        public int Sheath { get; set; } = 0;
+        public int RandomProperty { get; set; } = 0;
+        public int Block { get; set; } = 0;
+        public int ItemSet { get; set; } = 0;
+        public int MaxDurability { get; set; } = 0;
+        public int Area { get; set; } = 0;
+        public int Map { get; set; } = 0;
+        public int BagFamily { get; set; } = 0;
         public string ScriptName { get; set; }
-        public int DisenchantId { get; set; }
-        public int FoodType { get; set; }
-        public int MinMoneyLoot { get; set; }
-        public int MaxMoneyLoot { get; set; }
-        public int Duration { get; set; }
-        public int ExtraFlags { get; set; }
+        public int DisenchantId { get; set; } = 0;
+        public int FoodType { get; set; } = 0;
+        public int MinMoneyLoot { get; set; } = 0;
+        public int MaxMoneyLoot { get; set; } = 0;
+        public int Duration { get; set; } = 0;
+        public int ExtraFlags { get; set; } = 0;
     }
 }